+++ src/assets/css/landingPageInicio.css
@@ -5,7 +5,7 @@ h4 {
     font-weight: 400;
 }
 
-.LP-navbar {
+.LPNavbar {
     display: flex;
 	justify-content: space-between;
 	align-items: center;
@@ -16,16 +16,16 @@ h4 {
     background-color: rgb(0 0 0 / .8);; */
 }
 
-.LP-header {
+.LPHeader {
     /* width: 100%; */
     height: 100vh;
-    background-image: url('../img/LPInicio.png');
+    background-image: url('../img/LpInicio.png');
     background-repeat: no-repeat;
     background-size: 100%;
     /* margin: none; */
 }
 
-.LP-banner {
+.LPBanner {
     padding: 0 5%;
 }
 
@@ -45,7 +45,7 @@ button {
 	font-size: 1rem;
 }
 
-#banner-phrase {
+#bannerPhrase {
     color: #FFF;
     font-family: 'Montserrat', sans-serif;
     font-size: 64px;
