--- src/components/NoviLogo.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-import logo from "../assets/img/LogoNovi.png";
-
-function NoviLogo(){
-    return(
-        <div>
-            <img src={logo} alt="Logo Novi"/>
-        </div>
-    )
-}
-
-export default NoviLogo;
